requirements.txt
robotframework>=3.1
# robotframework-seleniumlibrary>=3.3.0
selenium>=3.4.0
